Complete the renaming of "geotiff" package into "earthobservation".
