+++ storage/sis-geotiff/src/test/java/org/apache/sis/test/suite/GeotiffTestSuite.java
@@ -30,8 +30,8 @@ import org.junit.BeforeClass;
  * @module
  */
 @Suite.SuiteClasses({
-    org.apache.sis.storage.geotiff.LandsatReaderTest.class,
-    org.apache.sis.storage.geotiff.ModisReaderTest.class
+    org.apache.sis.storage.earthobservation.LandsatReaderTest.class,
+    org.apache.sis.storage.earthobservation.ModisReaderTest.class
 })
 public final strictfp class GeotiffTestSuite extends TestSuite {
     /**
