+++ storage/sis-geotiff/src/test/resources/org/apache/sis/storage/earthobservation/Sample.txt
